% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test data}
\format{A data frame with 50 rows and 2 columns
\describe{
  \item{weight}{numeric, random values}
  \item{height}{numeric, random values}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
test_data
}
\description{
A dataset for testing roxygen2 for documenting a data frame
}
\keyword{datasets}
